search bar put in, feeds are personalized, admin.py added to forum, engine and accounts urls updated,

+++ biostar/forum/auth.py
@@ -1,6 +1,7 @@
 import datetime
 import logging
 
+from antispam import akismet
 from django.template import loader
 from django.conf import settings
 from django.contrib import messages
@@ -35,16 +36,19 @@ def get_votes(user, root):
     return store
 
 
-def create_subscription(post, user, sub_type=None):
+def create_subscription(post, user, sub_type=None, delete_exisiting=True):
     """
     Creates subscription to a post. Returns a list of subscriptions.
     """
-    # Drop all existing subscriptions for the user if these exists.
-    # Only one subscription may exist.
-    Subscription.objects.filter(post=post.root, user=user).delete()
-
-    # Create the subscription to the user.
-    Subscription.objects.create(post=post.root, user=user, type=sub_type)
+    # Drop all existing subscriptions for the user by default.
+    if delete_exisiting:
+        Subscription.objects.filter(post=post.root, user=user).delete()
+        # Create new subscription to the user.
+        Subscription.objects.create(post=post.root, user=user, type=sub_type)
+    # Update an existing subscription type.
+    else:
+        sub, created = Subscription.objects.get_or_create(post=post.root, user=user)
+        Subscription.objects.filter(pk=sub.pk).update(type=sub_type)
 
     # Recompute post subscription.
     subs_count = Subscription.objects.filter(post=post.root).exclude(type=Profile.NO_MESSAGES).count()
@@ -95,9 +99,9 @@ def post_tree(user, root):
             comment_tree.setdefault(post.parent_id, []).append(post)
         post.has_bookmark = int(post.id in bookmarks)
         post.has_upvote = int(post.id in upvotes)
-        post.can_accept = not post.is_toplevel and user == post.root.author
+        post.can_accept = not post.is_toplevel and (user == post.root.author or (user.is_authenticated and user.profile.is_moderator))
         post.can_moderate = user.is_authenticated and user.profile.is_moderator
-        post.is_editable = user.is_authenticated and (user == post.author or user.profile.is_moderator)
+        post.is_editable = user.is_authenticated and (user == post.author or (user.is_authenticated and user.profile.is_moderator))
         return post
 
     # Decorate the objects for easier access
@@ -128,6 +132,7 @@ def update_post_views(post, request, minutes=settings.POST_VIEW_MINUTES):
 
     # One view per time interval from each IP address.
     if not PostView.objects.filter(ip=ip, post=post, date__gt=since).exists():
+        # Update the last time
         PostView.objects.create(ip=ip, post=post, date=now)
         Post.objects.filter(pk=post.pk).update(view_count=F('view_count') + 1)
     return post
@@ -206,7 +211,7 @@ def delete_post(post, request):
     return url
 
 
-def moderate_post(request, action, post, comment=None, dupes=[], pid=None):
+def moderate_post(request, action, post, offtopic='', comment=None, dupes=[], pid=None):
     root = post.root
     user = request.user
     now = datetime.datetime.utcnow().replace(tzinfo=utc)
@@ -217,7 +222,7 @@ def moderate_post(request, action, post, comment=None, dupes=[], pid=None):
         messages.success(request, "Post bumped")
         return url
 
-    if action == MOD_OPEN:
+    if action == OPEN_POST:
         Post.objects.filter(uid=post.uid).update(status=Post.OPEN)
         messages.success(request, f"Opened post: {post.title}")
         return url
@@ -225,49 +230,36 @@ def moderate_post(request, action, post, comment=None, dupes=[], pid=None):
     if action == DELETE:
         return delete_post(post=post, request=request)
 
-    if action == TOGGLE_ACCEPT:
-        # Recompute accept count for post
-        change = -1 if post.has_accepted else + 1
-        Post.objects.filter(uid=post.uid).update(accept_count=F("accept_count") + change)
-        Post.objects.filter(uid=root.uid).update(accept_count=F("accept_count") + change)
-
-        return url
-
-    if action == MOVE_TO_ANSWER:
-        Post.objects.filter(uid=post.uid).update(type=Post.ANSWER, parent=post.root)
-        Post.objects.filter(uid=root.uid).update(reply_count=F("answer_count") + 1)
-        messages.success(request, "Moved comment to answer")
-        return url
-
-    if action == MOVE_TO_COMMENT or pid:
+    if pid:
         parent = Post.objects.filter(uid=pid).first() or post.root
         Post.objects.filter(uid=post.uid).update(type=Post.COMMENT, parent=parent)
         Post.objects.filter(uid=root.uid).update(reply_count=F("answer_count") - 1)
         messages.success(request, "Moved answer to comment")
         return url
 
-    if dupes:
-        # Change post status to closed.
-        Post.objects.filter(uid=post.uid).update(status=Post.CLOSED)
+    if offtopic:
         # Load comment explaining post closure.
+        tmpl = loader.get_template("messages/off_topic.md")
+        context = dict(user=post.author, comment=offtopic)
+        content = tmpl.render(context)
+
+        Post.objects.filter(uid=post.uid).update(status=Post.OFFTOPIC)
+        # Load answer explaining post being off topic.
+        Post.objects.create(content=content, type=Post.ANSWER, parent=post, author=user)
+        messages.success(request, "Marked the post as off topic.")
+
+        return url
+
+    if dupes:
+        # Load comment explaining post off topic label.
         tmpl = loader.get_template("messages/duplicate_posts.md")
         context = dict(user=post.author, dupes=dupes, comment=comment)
         content = tmpl.render(context)
-        # Create a comment with explanation as to why post is closed.
+
+        Post.objects.filter(uid=post.uid).update(status=Post.OFFTOPIC)
         Post.objects.create(content=content, type=Post.COMMENT, parent=post, author=user)
+        messages.success(request, "Closed duplicated post.")
+
         return url
 
-    messages.error(request, "Invalid moderation action given")
     return url
-
-
-def create_post(author, content, post_type, title="Title", tag_val="tag1, tag2", parent=None, root=None,):
-    "Used to create posts across apps"
-
-    post = Post.objects.create(title=title, content=content, tag_val=tag_val,
-                               author=author, type=post_type, parent=parent, root=root)
-
-    # Trigger notifications for subscribers and mentioned users
-    # async or synchronously
-
-    return post
